projetodesenvolve_api/requirements.txt
# requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
pandas==2.2.3
matplotlib==3.9.2
reportlab==4.2.5
python-dotenv==1.0.1
requests==2.32.3
django-cors-headers==4.6.0
gunicorn==23.0.0
setuptools<81.0.0